eshop/requirements/development.txt
-r common.txt

django-debug-toolbar

django-extensions
pygraphviz # for extensions
Werkzeug  # for extensions